Add duplicating logic in GA to the common package.
